Moving Curate::Ability into Ability class
We no longer need worry about the engine based-separation. So
consolidate to improve readability.
